pyproject.toml
[tool.poetry]
name = "api_aiohttp_mcc"
version = "0.1.0"
description = ""
authors = ["Kuzmich Valentine <asreto.asper@yahoo.com>"]

[tool.poetry.dependencies]
python = "^3.8"
aiohttp = "^3.7.4"
asyncpg = "^0.23.0"
PyYAML = "^5.4.1"
ujson = "^4.0.2"
loguru = "^0.5.3"
aiohttp-apispec = "^2.2.1"

[tool.poetry.dev-dependencies]
psycopg2 = "^2.8.6"
pandas = "^1.2.4"
flake8 = "^3.9.2"
pytest-aiohttp = "^0.3.0"
openpyxl = "^3.0.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
